+++ README.md
@@ -1,3 +1,13 @@
 # pizzapranks
 
 Wordpress theme for pizzapranks.com
+
+## Release notes
+
+### 3.2023 - Remove webpack/fancy JS
+
+Similar to CSS, also no longer need the transpilation or the hassle of keeping vendor dependencies updated
+
+### 1.2023 - Remove CSS buildsystem
+
+Want the current site owner to be able to make CSS changes, using the tooling makes this annoying/impossible
