+++ footer.php
@@ -3,7 +3,6 @@
 		Designed & Made with Love by <a href="https://www.twitter.com/thewatermethod" rel="designer">Matt Bevilacqua</a>, who owns <a href="https://whalingcityweb.com">Whaling City Web</a>.
 	</p>	  
 </footer>	
-		<?php wp_footer();?>
-	
+		<?php wp_footer();?>	
    </body>
 </html>
