+++ package.json
@@ -2,30 +2,11 @@
   "name": "pizzapranks",
   "version": "1.0.0",
   "description": "Wordpress theme for pizzapranks.com",
-  "main": "gulpfile.js",
-  "dependencies": {
-    "axios": "^0.21.1",
-    "moment": "^2.21.0",
-    "requirejs": "^2.3.5",
-    "vue": "^2.5.16",
-    "webfontloader": "^1.6.28"
-  },
-  "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "copy": "^0.3.1",
-    "path": "^0.12.7",
-    "stylelint": "^9.1.1",
-    "stylelint-config-standard": "^18.2.0",
-    "terser-webpack-plugin": "^3.0.3",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11"
-  },
+  "main": "js/main.js",
+  "dependencies": {},
+  "devDependencies": {},
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "gurp": "gulp -w",
-    "start": "webpack --mode=development -w",
-    "build": "webpack --mode=production"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
