+++ style.css
@@ -4,7 +4,7 @@ Theme URI: www.twilitgrotto.com/pisspranks
 Description: Custom Webcomic WP site
 Author: Matt Bevilacqua
 Author URI: http://twilitgrotto.com
-Version: 4.0
+Version: 4.2.0
 Tags: responsive
 
 License: GNU General Public License v2 or later
